django/requirements.txt
django
django-address
djangorestframework
markdown
django-filter
pyyaml
django-cors-headers

